(PC-35493)[API] chore: move offers function to offers_api
