+++ api/src/pcapi/core/offers/constants.py
@@ -1,0 +1,9 @@
+from pcapi.core.categories import subcategories
+
+ALLOWED_PRODUCT_SUBCATEGORIES = [
+    subcategories.SUPPORT_PHYSIQUE_MUSIQUE_CD.id,
+    subcategories.SUPPORT_PHYSIQUE_MUSIQUE_VINYLE.id,
+    subcategories.LIVRE_PAPIER.id,
+]
+
+
